% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_network_metrics.R
\name{summarize_network_metrics}
\alias{summarize_network_metrics}
\title{Summarize Directed Network List}
\usage{
summarize_network_metrics(net_list)
}
\arguments{
\item{net_list}{A list of directed adjacency matrices (can be from any source).}
}
\value{
A data frame with one row per network and the following columns:
\itemize{
\item net_id
\item n_nodes
\item n_edges
\item num_loops (number of unique feedback loops)
\item sigma_total (sum of SDs of in/out degrees)
\item node_overlap_score
\item avg_loop_size
}
}
\description{
Compute feedback loop and topology metrics for a list of directed networks.
}
