% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_topology.R
\name{summarize_topology}
\alias{summarize_topology}
\title{Summarize Network Topology}
\usage{
summarize_topology(adj_matrix, loops)
}
\arguments{
\item{adj_matrix}{A directed adjacency matrix.}

\item{loops}{A list of detected feedback loops.}
}
\value{
A list with total degree variability and number of loops.
}
\description{
Compute degree variability and loop stats from an adjacency matrix.
}
