% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_feedback_loops.R
\name{detect_feedback_loops}
\alias{detect_feedback_loops}
\title{Detect Feedback Loops}
\usage{
detect_feedback_loops(adj_matrix)
}
\arguments{
\item{adj_matrix}{A directed adjacency matrix.}
}
\value{
A list of cycles, each represented as a vector of node indices.
}
\description{
Detect cycles (feedback loops) in a directed adjacency matrix.
}
