% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_directed_networks.R
\name{generate_directed_networks}
\alias{generate_directed_networks}
\title{Generate All Directed Networks from a Skeleton (qgraph version)}
\usage{
generate_directed_networks(adj_matrix)
}
\arguments{
\item{adj_matrix}{A symmetric adjacency matrix representing the undirected skeleton.}
}
\value{
A list of adjacency matrices representing all possible directed networks.
}
\description{
Given an undirected adjacency matrix, generate all possible directed versions using qgraph.
}
