% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{plot_symptom_dynamics}
\alias{plot_symptom_dynamics}
\title{Plot Symptom Dynamics}
\usage{
plot_symptom_dynamics(S)
}
\arguments{
\item{S}{A matrix from simulate_dynamics (time x symptoms).}
}
\description{
Plots symptom intensities over time.
}
